Changed descriptions of the email part, in the view, so they make more sense to users.
As the gem stands right now, there's a select menu to choose the email part, with the descriptions "View HTML part" and "View plain-text part". The part *not* currently shown is the visible element in the menu on loading, so for example, if I'm viewing plaintext, the menu shows "View HTML part" on loading. However, this is awkward for two reasons:
1. When I select the menu and try to change to see the html part, I can't, because that's already selected
2. Select menu's usually show the currently selected option. In this case it's showing the element that's *not* currently selected (eg. on loading I see plaintext but it shows html)

I've changed this so that (on initial load) the element selected matches what's shown, and I've also tidied up the description of the select menu items.